# Make sizing computation much faster for nested containers