@penguinolog 2025-02-18
Super useful change.
please put `# fmt: off` / `# fmt: on` around enum - formatters love only 1 space around `=`.